feat(monorepo-2): execute foundation — @gt/core + turbo + 6 tools migrated
