+++ .gitignore
@@ -1,5 +1,9 @@
 # Based on https://raw.githubusercontent.com/github/gitignore/main/Node.gitignore
 
+# Turborepo
+.turbo
+packages/*/dist
+
 # Logs
 
 logs
