+++ packages/core/src/paths.client.test.ts
