+++ src/files-to-prompt/index.ts
@@ -3,9 +3,9 @@ import { mkdir, readdir, readFile } from "node:fs/promises";
 import { basename, dirname, extname, join, relative, resolve } from "node:path";
 import { logger, out } from "@app/logger";
 import { runTool } from "@app/utils/cli";
-import { formatBytes as _formatBytes } from "@app/utils/format";
 import { handleReadmeFlag } from "@app/utils/readme";
 import { estimateTokens, formatTokens } from "@ask/utils/helpers";
+import { formatBytes as _formatBytes } from "@gt/core/format";
 import type { FileSink } from "bun";
 import { Command } from "commander";
 import { minimatch } from "minimatch";
