+++ src/json/index.ts
@@ -2,8 +2,8 @@ import { existsSync, readFileSync } from "node:fs";
 import { resolve } from "node:path";
 import { out } from "@app/logger";
 import { runTool } from "@app/utils/cli";
-import { SafeJSON } from "@app/utils/json";
 import { handleReadmeFlag } from "@app/utils/readme";
+import { SafeJSON } from "@gt/core/json";
 import { Command } from "commander";
 import { fromToon as decode, toToon as encode } from "./lib/toon";
