+++ src/npm-package-diff/index.ts
@@ -6,8 +6,8 @@ import path from "node:path";
 import { out } from "@app/logger";
 import { resolvePathWithTilde } from "@app/utils";
 import { isVerbose, runTool } from "@app/utils/cli";
-import { SafeJSON } from "@app/utils/json";
 import { handleReadmeFlag } from "@app/utils/readme";
+import { SafeJSON } from "@gt/core/json";
 import boxen from "boxen";
 import chalk from "chalk";
 import chokidar, { type FSWatcher } from "chokidar";
