+++ src/usage/index.ts
@@ -3,8 +3,8 @@ import { UsageDatabase } from "@app/ask/output/UsageDatabase";
 import { dynamicPricingManager } from "@app/ask/providers/DynamicPricing";
 import { logger, out } from "@app/logger";
 import { runTool } from "@app/utils/cli";
-import { formatDateTime } from "@app/utils/date";
 import { SafeJSON } from "@app/utils/json";
+import { formatDateTime } from "@gt/core/date";
 import chalk from "chalk";
 import Table from "cli-table3";
 import { Command } from "commander";
