+++ src/utils/format.ts
@@ -1,415 +1,1 @@
-/**
- * Shared formatting utilities for CLI tools.
- * Consolidates formatDuration, formatBytes, formatCost, formatTokens,
- * formatRelativeTime, formatList, and formatNumber from across the codebase.
- */
-
-// ============= Duration Formatting =============
-
-type DurationUnit = "ms" | "s" | "min";
-type DurationStyle = "tiered" | "hm-always" | "hm-smart" | "hms";
-
-/**
- * Format a duration to a human-readable string.
- *
- * @param value - The numeric duration value
- * @param unit - The unit of the input value: "ms" (default), "s", or "min"
- * @param style - Output format style:
- *   - "tiered" (default): ms → s → m+s → h+m (from colors.ts)
- *   - "hm-always": Always "Xh Ym" (from timely/date.ts)
- *   - "hm-smart": Omits zeros, "< 1m" for sub-minute, rounds minutes (from history.ts)
- *   - "hms": Includes seconds at all levels (from ask/cli.ts)
- */
-export function formatDuration(value: number, unit: DurationUnit = "ms", style: DurationStyle = "tiered"): string {
-    // Normalize to milliseconds
-    let ms: number;
-    switch (unit) {
-        case "s":
-            ms = value * 1000;
-            break;
-        case "min":
-            ms = value * 60000;
-            break;
-        default:
-            ms = value;
-    }
-
-    const totalSeconds = Math.floor(ms / 1000);
-    const totalMinutes = ms / 60000;
-
-    switch (style) {
-        case "tiered": {
-            if (ms < 1000) {
-                return `${Math.round(ms)}ms`;
-            }
-            if (ms < 60000) {
-                return `${(ms / 1000).toFixed(1)}s`;
-            }
-            if (ms < 3600000) {
-                return `${Math.floor(ms / 60000)}m ${Math.floor((ms % 60000) / 1000)}s`;
-            }
-            return `${Math.floor(ms / 3600000)}h ${Math.floor((ms % 3600000) / 60000)}m`;
-        }
-
-        case "hm-always": {
-            const hours = Math.floor(totalSeconds / 3600);
-            const minutes = Math.floor((totalSeconds % 3600) / 60);
-            return `${hours}h ${minutes}m`;
-        }
-
-        case "hm-smart": {
-            if (totalMinutes < 1) {
-                return "< 1m";
-            }
-            let hours = Math.floor(totalMinutes / 60);
-            let mins = Math.round(totalMinutes % 60);
-            if (mins === 60) {
-                hours++;
-                mins = 0;
-            }
-            if (hours === 0) {
-                return `${mins}m`;
-            }
-            if (mins === 0) {
-                return `${hours}h`;
-            }
-            return `${hours}h ${mins}m`;
-        }
-
-        case "hms": {
-            const seconds = Math.floor(ms / 1000);
-            const minutes = Math.floor(seconds / 60);
-            const hours = Math.floor(minutes / 60);
-            if (hours > 0) {
-                return `${hours}h ${minutes % 60}m ${seconds % 60}s`;
-            }
-            if (minutes > 0) {
-                return `${minutes}m ${seconds % 60}s`;
-            }
-            return `${seconds}s`;
-        }
-
-        default: {
-            // Exhaustiveness check — if a new `style` is added without a case
-            // arm, `satisfies never` becomes a compile error here. No variable
-            // is allocated (dashboard tsconfig has noUnusedLocals: true).
-            style satisfies never;
-            return `${ms}ms`;
-        }
-    }
-}
-
-// ============= Duration Parsing =============
-
-// Single value + unit, e.g. "30m", "2hours", "90s"
-const SIMPLE_DURATION_PATTERN = /^(\d+)\s*(s|sec|secs|seconds?|m|min|mins|minutes?|h|hr|hrs|hours?)$/i;
-
-// Multi-part durations, e.g. "1h30m", "2h15min10s". Each group is optional,
-// allowing any combination of hours/minutes/seconds:
-//   (?:(\d+)\s*h...)? — optional hours  (capture group 1)
-//   (?:(\d+)\s*m...)? — optional minutes (capture group 2)
-//   (?:(\d+)\s*s...)? — optional seconds (capture group 3)
-// Unit suffixes accept common spellings: h/hr/hrs/hour/hours, m/min/mins/minute/minutes, s/sec/secs/second/seconds
-const COMPOUND_DURATION_PATTERN =
-    /^(?:(\d+)\s*h(?:ours?|rs?|r)?)?\s*(?:(\d+)\s*m(?:in(?:utes?|s)?)?)?\s*(?:(\d+)\s*s(?:ec(?:onds?|s)?)?)?$/i;
-
-/**
- * Parse a human-readable duration string into milliseconds.
- *
- * Supports:
- *   - Simple: "30s", "30m", "30min", "30minutes", "2h", "2hours"
- *   - Compound: "1h30m", "2h15min", "20m 1s", "1h30m15s"
- *   - Plain number: treated as minutes (e.g. "30" → 30 minutes)
- *
- * Returns 0 if the input cannot be parsed.
- */
-export function parseDuration(input: string): number {
-    const trimmed = input.trim();
-
-    if (/^\d+$/.test(trimmed)) {
-        return Number.parseInt(trimmed, 10) * 60 * 1000;
-    }
-
-    const simple = trimmed.match(SIMPLE_DURATION_PATTERN);
-
-    if (simple) {
-        const value = Number.parseInt(simple[1], 10);
-        const unit = simple[2].toLowerCase();
-
-        if (unit.startsWith("s")) {
-            return value * 1000;
-        }
-
-        if (unit.startsWith("m")) {
-            return value * 60 * 1000;
-        }
-
-        if (unit.startsWith("h")) {
-            return value * 60 * 60 * 1000;
-        }
-    }
-
-    const compound = trimmed.match(COMPOUND_DURATION_PATTERN);
-
-    if (compound && (compound[1] || compound[2] || compound[3])) {
-        const hours = compound[1] ? Number.parseInt(compound[1], 10) : 0;
-        const minutes = compound[2] ? Number.parseInt(compound[2], 10) : 0;
-        const seconds = compound[3] ? Number.parseInt(compound[3], 10) : 0;
-        return (hours * 3600 + minutes * 60 + seconds) * 1000;
-    }
-
-    return 0;
-}
-
-// ============= Relative Time =============
-
-interface FormatRelativeTimeOptions {
-    /** Maximum days before falling back to absolute date. Default: Infinity */
-    maxDays?: number;
-    /** Fallback formatter when age exceeds maxDays */
-    fallbackFormat?: (date: Date) => string;
-    /** Rounding mode for time units. Default: "floor" */
-    rounding?: "floor" | "round";
-    /** Compact output: "5m ago", "3h ago", "2d ago". Default: false */
-    compact?: boolean;
-}
-
-/**
- * Format a Date as a relative time string like "5 minutes ago".
- * With compact: true, returns "5m ago", "3h ago", "2d ago".
- *
- * @param date - The date to format
- * @param options - Formatting options
- */
-export function formatRelativeTime(date: Date, options?: FormatRelativeTimeOptions): string {
-    const { maxDays = Infinity, fallbackFormat, rounding = "floor", compact = false } = options ?? {};
-    const round = rounding === "round" ? Math.round : Math.floor;
-
-    const diffMs = Date.now() - date.getTime();
-    const diffMinutes = round(diffMs / 60000);
-    const diffHours = round(diffMs / 3600000);
-    const diffDays = round(diffMs / 86400000);
-
-    if (maxDays !== Infinity && diffDays >= maxDays && fallbackFormat) {
-        return fallbackFormat(date);
-    }
-
-    if (compact) {
-        if (diffMinutes < 1) {
-            return "now";
-        }
-        if (diffHours < 1) {
-            return `${diffMinutes}m ago`;
-        }
-        if (diffDays < 1) {
-            return `${diffHours}h ago`;
-        }
-        if (diffDays < 7) {
-            return `${diffDays}d ago`;
-        }
-        return date.toISOString().slice(0, 10);
-    }
-
-    if (diffMinutes < 1) {
-        return "just now";
-    }
-    if (diffMinutes === 1) {
-        return "1 minute ago";
-    }
-    if (diffMinutes < 60) {
-        return `${diffMinutes} minute${diffMinutes !== 1 ? "s" : ""} ago`;
-    }
-    if (diffHours === 1) {
-        return "1 hour ago";
-    }
-    if (diffHours < 24) {
-        return `${diffHours} hour${diffHours !== 1 ? "s" : ""} ago`;
-    }
-    return `${diffDays} day${diffDays === 1 ? "" : "s"} ago`;
-}
-
-/**
- * Compact "time since last message" for live dashboards.
- * Examples: `5s ago`, `2m 15s ago`, `1h 5m 3s ago`
- */
-export function formatLastMessageAgo(deltaMs: number): string {
-    const totalSeconds = Math.max(0, Math.floor(deltaMs / 1000));
-    const hours = Math.floor(totalSeconds / 3600);
-    const minutes = Math.floor((totalSeconds % 3600) / 60);
-    const seconds = totalSeconds % 60;
-
-    if (hours > 0) {
-        return `${hours}h ${minutes}m ${seconds}s ago`;
-    }
-
-    if (minutes > 0) {
-        return `${minutes}m ${seconds}s ago`;
-    }
-
-    return `${seconds}s ago`;
-}
-
-/**
- * Compact elapsed duration with day/hour/minute parts (no seconds).
- * Examples: `0m`, `22m`, `1h 22m`, `1d 1h 22m`
- */
-export function formatElapsedDhM(deltaMs: number): string {
-    const totalMinutes = Math.floor(Math.max(0, deltaMs) / 60000);
-    const days = Math.floor(totalMinutes / (60 * 24));
-    const hours = Math.floor((totalMinutes % (60 * 24)) / 60);
-    const minutes = totalMinutes % 60;
-    const parts: string[] = [];
-
-    if (days > 0) {
-        parts.push(`${days}d`);
-    }
-
-    if (hours > 0) {
-        parts.push(`${hours}h`);
-    }
-
-    if (minutes > 0 || parts.length === 0) {
-        parts.push(`${minutes}m`);
-    }
-
-    return parts.join(" ");
-}
-
-// ============= Bytes =============
-
-/**
- * Format bytes to human-readable size string.
- * Sub-KB values show no decimals; larger values show 1 decimal place.
- */
-export function formatBytes(bytes: number): string {
-    const units = ["B", "KB", "MB", "GB", "TB"];
-    if (bytes < 1024) {
-        return `${bytes} B`;
-    }
-    let size = bytes;
-    let i = 0;
-    while (size >= 1024 && i < units.length - 1) {
-        size /= 1024;
-        i++;
-    }
-    return `${size.toFixed(1)} ${units[i]}`;
-}
-
-// ============= Cost & Tokens =============
-
-/**
- * Format a cost value as "$X.XXXX" (4 decimal places).
- */
-export function formatCost(cost: number): string {
-    return `$${cost.toFixed(4)}`;
-}
-
-/**
- * Format a token count with K/M abbreviations.
- */
-export function formatTokens(tokens: number): string {
-    if (tokens >= 1_000_000) {
-        return `${(tokens / 1_000_000).toFixed(1)}M`;
-    }
-    if (tokens >= 1_000) {
-        return `${(tokens / 1_000).toFixed(1)}K`;
-    }
-    return tokens.toString();
-}
-
-// ============= List =============
-
-/**
- * Format a list of items, truncating if too long.
- */
-export function formatList(items: string[], maxShow = 5): string {
-    if (items.length <= maxShow) {
-        return items.join(", ");
-    }
-    const shown = items.slice(0, maxShow);
-    const remaining = items.length - maxShow;
-    return `${shown.join(", ")} +${remaining} more`;
-}
-
-// ============= Number =============
-
-/**
- * Format a large number with K/M/B abbreviations.
- */
-export function formatNumber(n: number): string {
-    if (n >= 1_000_000_000) {
-        return `${(n / 1_000_000_000).toFixed(1)}B`;
-    }
-    if (n >= 1_000_000) {
-        return `${(n / 1_000_000).toFixed(1)}M`;
-    }
-    if (n >= 1_000) {
-        return `${(n / 1_000).toFixed(1)}K`;
-    }
-    return n.toString();
-}
-
-// ============= Timestamps =============
-
-/**
- * Format current time as HH:MM:SS.mmm (24h with milliseconds).
- * Useful for log prefixes where you need to correlate wall-clock time.
- */
-export function formatTimestamp(date: Date = new Date()): string {
-    const h = String(date.getHours()).padStart(2, "0");
-    const m = String(date.getMinutes()).padStart(2, "0");
-    const s = String(date.getSeconds()).padStart(2, "0");
-    const ms = String(date.getMilliseconds()).padStart(3, "0");
-    return `${h}:${m}:${s}.${ms}`;
-}
-
-/**
- * Locale clock string, always 24-hour. The single source of display-time
- * formatting for UIs (dashboard etc.) so `hour12:false` isn't repeated at
- * every call site. Pure — safe to import in the browser.
- *
- * @example formatClock(ts)                       // "14:05"
- * @example formatClock(ts, { seconds: true })    // "14:05:09"
- * @example formatClock(ts, { date: "numeric" })  // "5/15, 14:05"
- * @example formatClock(ts, { date: "short" })    // "May 15, 14:05"
- */
-export function formatClock(
-    input: Date | string | number,
-    opts: { seconds?: boolean; date?: "none" | "numeric" | "short" } = {}
-): string {
-    const d = input instanceof Date ? input : new Date(input);
-    if (Number.isNaN(d.getTime())) {
-        return String(input);
-    }
-
-    const time = d.toLocaleTimeString([], {
-        hour: "2-digit",
-        minute: "2-digit",
-        ...(opts.seconds ? { second: "2-digit" } : {}),
-        hour12: false,
-    });
-
-    if (!opts.date || opts.date === "none") {
-        return time;
-    }
-
-    const datePart = d.toLocaleDateString(
-        [],
-        opts.date === "short" ? { month: "short", day: "numeric" } : { month: "numeric", day: "numeric" }
-    );
-    return `${datePart}, ${time}`;
-}
-
-/**
- * Create a simple stopwatch function that returns formatted elapsed time.
- * For a full-featured stopwatch with laps/stamps, use Stopwatch class from @app/utils/Stopwatch.
- *
- * @example
- * const elapsed = createStopwatch();
- * // ... do work ...
- * console.log(elapsed()); // "1.5s"
- */
-export function createStopwatch(): () => string {
-    const start = performance.now();
-    return () => formatDuration(performance.now() - start);
-}
+export * from "@gt/core/format";
